feat: add isProcessing

+++ src/components/MainTweet.vue
@@ -35,7 +35,7 @@
                 <img src="../assets/images/tweet_reply.svg" alt="">
             </div>
         
-            <div v-if="!tweet.isLiked" @click.stop.prevent="addLike(tweet.id)" >
+            <div v-if="!tweet.isLiked" @click.stop.prevent="addLike(tweet.id)">
                 <img src="../assets/images/tweet_like.svg" alt="">
             </div>
             <div v-else @click.stop.prevent="deleteLike(tweet.id)" >
@@ -69,7 +69,8 @@ export default {
         return {
             tweet: this.initialData,
             isModalToggled: this.iniIsModalToggled,
-            isLoading: true
+            isLoading: true,
+            isProcessing: false
         }
     },
     watch: {
@@ -90,6 +91,9 @@ export default {
         },
          async addLike(tweet_id){
             try {
+                if (this.isProcessing) return
+                this.isProcessing = true
+
                 const response = await tweetsAPI.addLike({tweet_id})
                 console.log(response)
                 this.tweet = {
@@ -97,9 +101,11 @@ export default {
                     isLiked: true,
                     likeCounts: this.tweet.likeCounts + 1
                 }
+                this.isProcessing = false
             }
             catch (error) {
                 console.log(error.message)
+                this.isProcessing = false
                 Toast.fire({
                     icon: 'error',
                     title: `目前無法操作,請稍後再試`,
@@ -108,6 +114,9 @@ export default {
         },
         async deleteLike(tweet_id){
             try {
+                if (this.isProcessing) return
+                this.isProcessing = true
+
                 const response = await tweetsAPI.deleteLike({tweet_id})
                 console.log(response)
                 this.tweet = {
@@ -115,9 +124,11 @@ export default {
                     isLiked: false,
                     likeCounts: this.tweet.likeCounts - 1
                 }
+                this.isProcessing = false
             }
             catch (error) {
                 console.log(error.message)
+                this.isProcessing = false
                 Toast.fire({
                     icon: 'error',
                     title: `目前無法操作,請稍後再試`,
