+++ src/components/RecommendUsers.vue
@@ -4,6 +4,7 @@
         <div class="recommend__container">
 
             <router-link :to="{ name: 'user-detail', params: {id: user.id}, query: {tab: 'tweet'}}" v-for="user in users" :key="user.id" class="recommend__user">
+            <!-- 個人資訊區 -->
                 <div class="recommend__user--avatar">
                     <img :src="user.profilePhoto" alt="">
                 </div>
@@ -11,9 +12,14 @@
                     <p>{{user.name}}</p>
                     <span>@{{user.account}}</span>
                 </div>
-
-                <button @click.stop.prevent="addFollowship(user.id)" v-if="!user.isFollowed" class="prim-button prim-button__unfollowed">跟隨</button>
-                <button @click.stop.prevent="deleteFollowship(user.id)" v-else class="prim-button prim-button__followed">正在跟隨</button>
+            <!-- 按鈕區 -->
+                <button 
+                @click.stop.prevent="addFollowship(user.id)" 
+                v-if="!user.isFollowed" 
+                class="primbutton primbutton__unfollowed" :class="{primbutton__unfollowed_processing: user.isProcessing}">跟隨</button>
+                <button 
+                @click.stop.prevent="deleteFollowship(user.id)" v-else 
+                class="primbutton primbutton__followed" :class="{primbutton__followed_processing: user.isProcessing}">正在跟隨</button>
             </router-link>
 
         </div>
@@ -28,20 +34,26 @@ import { mapState } from 'vuex'
 export default {
     data () {
         return {
-            users: []
+            users: [],
         }
     },
     created () {
         this.fecthTopUsers()
     },
     computed: {
-        ...mapState(['currentUser', 'isAuthenticated'])
+        ...mapState(['currentUser', 'isAuthenticated']),
     },
     methods: {
         async fecthTopUsers () {
             try {
                 const { data } = await followshipsAPI.getTopUsers()
-                this.users = data.filter(data => data.id !== this.currentUser.id).slice(0, 10)
+                this.users = data
+                .filter(data => data.id !== this.currentUser.id).slice(0, 10)
+                .map((user) => {
+                    return {
+                        ...user, isProcessing: false
+                    }
+                })
             }
             catch (error) {
                 console.log(error)
@@ -53,8 +65,12 @@ export default {
         },
         async addFollowship (id) {
             try{
+                //該user的請求處理中就return
+                if (this.checkProcessing (id)) return
+                //將該user變成處理中
+                this.changeIsProcessing (id)
+                
                 const response = await followshipsAPI.addFollowship({id})
-                console.log(response)
 
                 if (response.status !==  200) {
                     throw new Error(response.data.message)
@@ -68,10 +84,11 @@ export default {
                     }
                     return user
                 })
-                this.$emit('after-like-user', this.users)
+                this.changeIsProcessing (id)
             }
             catch(error) {
                 const message = error.response.data.message
+                this.changeIsProcessing (id)
                 console.log(message)
                 return Toast.fire({
                     icon: 'error',
@@ -82,6 +99,9 @@ export default {
         },
         async deleteFollowship (followingId) {
             try{
+                if (this.checkProcessing (followingId)) return
+                this.changeIsProcessing (followingId)
+
                 const response = await followshipsAPI.deleteFollowship({followingId})
 
                 if (response.status !==  200) {
@@ -97,9 +117,11 @@ export default {
                     }
                     return user
                 })
+                this.changeIsProcessing (followingId)
             }
             catch(error) {
                 const message = error.response.data.message
+                this.changeIsProcessing (followingId)
                 console.log(message)
                 return Toast.fire({
                     icon: 'error',
@@ -107,6 +129,21 @@ export default {
                 })
             }
         },
-    }
+        //切換選中user的isProcessing
+        changeIsProcessing (userId) {
+            this.users = this.users.map(user => {
+                if (userId === user.id) {
+                    return { ...user, isProcessing: !user.isProcessing }
+                }
+                return { ...user }
+            })
+        },
+        //確認選中user是否isProcessing
+        checkProcessing (userId) {
+            const clickedUser = this.users.find(user => user.id === userId)
+            if (clickedUser.isProcessing) return true
+            return false
+        }
+    }   
 }
 </script>
